Move cypress base to <root> lerna modules
