+++ modules/cypress-base/src/commands/index.ts
@@ -17,4 +17,7 @@
  *
  */
 
+/// <reference types="cypress" />
+/// <reference path="../../types/index.d.ts" />
+
 import "./dom";
