+++ modules/cypress-base/src/constants/index.ts
