+++ modules/cypress-base/src/constants/login-page-dom-constants.ts
@@ -17,7 +17,10 @@
  *
  */
 
-export class HeaderDomConstants {
+/**
+ * Class containing Login Page DOM constants.
+ */
+export class LoginPageDomConstants {
 
     /**
      * Private constructor to avoid object instantiation from outside
@@ -25,8 +28,10 @@ export class HeaderDomConstants {
      *
      * @hideconstructor
      */
+    // eslint-disable-next-line @typescript-eslint/no-empty-function
     private constructor() { }
-    
-    public static readonly AVATAR_ICON_DATA_ATTR: string = "app-header-user-avatar";
-    public static readonly LOGOUT_BUTTON_DATA_ATTR: string = "app-header-dropdown-link";
+
+    public static readonly USERNAME_INPUT_DATA_ATTR: string = "#usernameUserInput";
+    public static readonly PASSWORD_INPUT_DATA_ATTR: string = "#password";
+    public static readonly CONTINUE_BUTTON_DATA_ATTR: string = "Continue";
 }
