+++ modules/cypress-base/tsconfig.json
@@ -1,0 +1,21 @@
+{
+    "compilerOptions": {
+        "declaration": true,
+        "experimentalDecorators": true,
+        "lib": [
+            "es5",
+            "dom"
+        ],
+        "outDir": "dist",
+        "rootDir": "src",
+        "target": "es5",
+        "types": [
+            "cypress",
+            "node"
+        ]
+    },
+    "include": [
+        "src"
+    ],
+    "extends": "../../configs/common.tsconfig.json"
+}
