--- tests/modules/cypress-base/tsconfig.json
@@ -1,3 +1,0 @@
-{
-    "extends": "../../tsconfig"
-}
